分类拖拽代码

+++ mall-product/src/main/java/com/blackman/mall/product/service/CategoryService.java
@@ -20,6 +20,6 @@ public interface CategoryService extends IService<CategoryEntity> {
 
     List<CategoryEntity> queryCategoryTree();
 
-    void removeMenusById(Integer[] ids);
+    void removeMenusById(Long[] ids);
 }
